// src/app/core/models/auth-info.ts
export interface AuthInfo {
    access_token: string;
    token_type: string;
    expires_in: number;
    email: string;
    first_name: string;
    last_name: string;
    mastercustomerid: string;
    user_id: string;
    user_type: string;
    speciality: string;
    zip: string;
    phone_number: string;
    country: string;
    status: string;
    expiry: string;
}